services/notifications/requirements.txt
mysql-connector-python==8.0.26
pika==1.2.0
boto3==1.20.2
botocore==1.23.2
